+++ +sbfsem/+core/VesselAdjacency.m
@@ -13,9 +13,9 @@ classdef VesselAdjacency < sbfsem.core.StructureAPI
 			obj@sbfsem.core.StructureAPI(ID, source);
 
 			if nargin < 3
-				obj.transform = sbfsem.core.Transforms.Viking;
+				obj.transform = sbfsem.builtin.Transforms.Standard;
 			else
-				obj.transform = sbfsem.core.Transforms.fromStr(transform);
+				obj.transform = sbfsem.builtin.Transforms.fromStr(transform);
 			end
 
 			% Instantiate OData client